Moved ProjectResetter to the projects package
